refactor: splits type definitions in separate files
so it's easier to move them closer to where they have their actual interface responsibility
