+++ src/labeler-yml/generate.test.ts
@@ -3,7 +3,7 @@ import { readFileSync } from "node:fs";
 import { EOL } from "node:os";
 import { describe, it } from "node:test";
 import { parse as parseYaml } from "yaml";
-import type { IVirtualCodeOwnersCST } from "../../types/types.js";
+import type { IVirtualCodeOwnersCST } from "../../types/virtual-code-owners-cst.js";
 import generateLabelerYml from "./generate.js";
 import readTeamMap from "../team-map/read.js";
 import readVirtualCodeOwners from "../virtual-code-owners/read.js";
