+++ src/labeler-yml/generate.ts
@@ -1,9 +1,9 @@
 import { EOL } from "node:os";
+import type { ITeamMap } from "types/team-map.js";
 import type {
   IInterestingCSTLine,
-  ITeamMap,
   IVirtualCodeOwnersCST,
-} from "types/types.js";
+} from "types/virtual-code-owners-cst.js";
 
 const DEFAULT_WARNING =
   `#${EOL}` +
