+++ src/virtual-code-owners/parse.ts
@@ -1,11 +1,11 @@
 import { EOL } from "node:os";
+import type { ITeamMap } from "types/team-map.js";
 import type {
-  ITeamMap,
   IUser,
   IVirtualCodeOwnerLine,
   IVirtualCodeOwnersCST,
   UserType,
-} from "types/types.js";
+} from "types/virtual-code-owners-cst.js";
 import { isEmailIshUsername } from "../utensils.js";
 
 export function parse(
