+++ types/main.d.ts
@@ -1,0 +1,8 @@
+export interface IOptions {
+  virtualCodeOwners: string;
+  virtualTeams: string;
+  codeOwners: string;
+  emitLabeler: boolean;
+  labelerLocation: string;
+  dryRun: boolean;
+}
