+++ types/team-map.d.ts
@@ -1,0 +1,3 @@
+export interface ITeamMap {
+  [teamName: string]: string[];
+}
